+++ bee-ternary/src/t3b1.rs
@@ -172,4 +172,8 @@ impl RawEncodingBuf for T3B1Buf {
     fn as_slice_mut(&mut self) -> &mut Self::Slice {
         unsafe { &mut *(Self::Slice::make(self.0.as_ptr() as _, 0, self.1) as *mut _) }
     }
+
+    fn capacity(&self) -> usize {
+        self.0.capacity()
+    }
 }